Time: 67 ms (24.86%) | Memory: 59.2 MB (74.95%) - LeetSync

+++ 2665-minimum-time-to-repair-cars/minimum-time-to-repair-cars.cpp
@@ -6,7 +6,7 @@ public:
         auto f =[&](long long mid){
             long long car =0;
             for(auto x : ranks){
-                car+=(int)sqrtl(mid/x);
+                car+=sqrt(mid/x);
             }
             return car>=cars;
         };
